mathml/relations/html5-tree/class-2.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Class</title>
<link rel="help" href="http://www.mathml-association.org/MathMLinHTML5/S2.html#SS2">
<meta name="assert" content="Verify whether the getElementsByClassName() works for MathML elements.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  var test_id = async_test("Test getElementsByClassName()");
  window.addEventListener("DOMContentLoaded", function() {
    var mtext = document.getElementsByClassName("cl");
    test_id.step(function() {
      assert_equals(mtext.length, 3);
      var mtext_ref = document.body.firstElementChild.firstElementChild;
      mtext_ref = mtext_ref.nextElementSibling.nextElementSibling
      assert_equals(mtext[0], mtext_ref);
      mtext_ref = mtext_ref.nextElementSibling.nextElementSibling;
      assert_equals(mtext[1], mtext_ref);
      mtext_ref = mtext_ref.nextElementSibling.nextElementSibling;
      assert_equals(mtext[2], mtext_ref);
    });
    test_id.done();
  });
</script>
</head>
<body>
  <math>
    <mtext class="cl_"></mtext>
    <mtext class="c"></mtext>
    <mtext class="cl"></mtext>
    <mtext class="cl_"></mtext>
    <mtext class="cl"></mtext>
    <mtext class="c"></mtext>
    <mtext class="cl"></mtext>
    <mtext class="cl_"></mtext>
  </math>
</body>
</html>
